Overhaul for `stream_to`.
I realised I'd left `stream_to` in a bit of an awkward state for
@dpriedel to build on.  He wrote up a patch to use parameter packs
instead of tuples, which is a lot friendlier, and it was a lot faster
too.

In this change I rearrange things to make that easier, and optimise
out the endless buffer allocations.  Also, just because we can, it's
now possible to insert a row in the form of a container, nust just
as a `std::tuple`.  Extending it to support parameter packs should be
a doddle.

Oh, as a bonus, I optimised it to skip the escaping step (which
involves an extra buffer copy) on types which just don't need any
escaping, such as integral types.

And then there are the tests...  We had tests, but they really didn't
do much actual testing of the data they inserted.  I added some.  And
along the way I found that the `ipv4` test type never converted to
strings properly.