Fix test breakage on older gcc/clang.
No idea why this doesn't show up on newer compilers.  It's not in any
way an issue otherwise, but it seems that the `ipv4` type (defined only
in test code) under-budgets the buffer space that its calls to convert
the octets demand.  They won't actually use that space, so no harm
done, except the per-octet conversion will complain.