Check for sufficient `std::variant` support.
On my laptop running Ubuntu 18.04 with all packaged gcc and clang
versions installed, and a few newer ones up to gcc 9 and clang 8,
clang 6 breaks in `std::variant` on its base class being private.
On my desktop urnning Ubuntu 19.10 with a similar compiler setup,
for some reason I see the same problem with gcc 7 as well.

So while we still support those compilers, check that we can work
with the available `std:;variant` and dropp support if not.

The code does still assume that we have `<variant>`.  This is not a
workaround for compilers without C++17.