New `stream_to` API, thanks to @dpriedel!
Dave Riedel has contributed a new, faster, more convenient way to feed
rows of data into `stream_to`.  A new method simply accepts variable
arguments, which all become fields in a single row.

Along the way Dave has done some performance measurements and found a
very nice speedup, even compared to the `std::tuple` function.  In all
honesty I hadn't been expecting that — I thought an optimising
compiler would end up producing the same code.

So thanks, Dave, for a nice and productive collaboration.  It's not
just this extra function either.  A bunch of other code ended up nicer
as a result of the back-and forth, and support for `std::variant` was
one of the byproducts.

I'll do a bit more tweaking on this, but I think the principle will
hold up nicely.