Parallelise `test_all.py`.
I hope this holds up under pressure.  The "broad-spectrum" tests in this
script were taking too long, and to some extent, without reason.

You see, each build breaks down into 3 phases: configure, build, test.
Of those, the first is single-threaded and we could run a whole lot of
those phases in parallel.  They could also run concurrently with the
build and test stages.  The build stage uses up all the CPUs it can get
its hands on, so we can only run one of those at a time.  Then the test
stage is (currently) not concurrency-safe, so we can run only one of
those at a time — but it's fine to overlap it with the other phases.
The more the better, in fact.  If test runs are serial, they form a
bottleneck for a part of the overall build time.

The script is getting awfully big, and that's something I hate.  One of
the causes is the asymmetry of having one stage which calls for as many
worker processes as CPUs, as well as two which call for one worker each.
These two categories are managed very differently, using callbacks and
`multiprocessing.pool.Pool` for one, and `multiprocessing.Queue` for the
other.

And that's not complex enough.  I may still throw in `asyncio` for good
measure.  It'd be great to be able to report errors a little faster.