Set CMake `CMP0074` policy.
See discussion in #306.

Unfortunately my development systems running Ubuntu 12.04 still have
CMake 3.7, which does not recognise the policy for enabling
`PostgreSQL_ROOT` as a means of controlling the paths to libpq headers
and library.  So I have to check the CMake version before enabling the
feature.  It feels like it sort of defeats the purpose of having these
forward-compatible policies.