Work around compiler warning.
Hard to find a way of saying "these fields may not actually be needed"
in a way that all my compilers will accept without issuing a warning.

In this case, clang 6 did not accept `[[maybe_unused]]` on a data
member.  We'll see if just inlining the data member and leaving out
the attribute will do the job.