Fix underestimated buffer size for floats.
Addresses #328, though we'll need review before we can close that one.

My calculation of buffer sizes needed to represent floating-point values
as strings was based on a misunderstanding of the C++ standard.  It
became a problem when `to_chars` is available for float types.

The bad news is that this larger buffer size will break the Small String
Optimisation for `double` values.  So expect a slowdown for `double`.

There is some minor good news to help make up for this: I was budgeting
one unnecessary extra byte in `to_string`, and that's been fixed.  The
change broke tests because the `bytea` conversion accidentally
_underestimated_ the buffer size required.  So you may actually get the
short string optimisation in some cases where previously you didn't.