Smoke testing for #328.
This is not great.  But it'll help guard against regressions or variants
of #328 in that critical period where some compilers build libpqxx with
floating-point `to_chars` out of the box, but my compiler doesn't
support it.  And probably overlapping with that, the period where
different `to_chars` implementations produce slightly different outputs.

Once we're through that, we can probably clean this up and assume that
`to_chars` just works.  And we'll be able to ditch a lot of complexity
in the string conversion code itself as well.