Rewrite `demangle_type_name` yet again.
Using `std::unique_ptr` to ensure freeing of the `__cxa_demangle` buffer
was a nice idea, but it led to a confusing gcc warning about `noexcept`.
Basically the compiler complained that I was passing it a `std::free()`
callback, even though the callback should be `noexcept`.

The compiler noted that `std::free` never throws, and so suggested
making the callback `noexcept`.  Great, except that it's implementation
specific.  behaviour.  The C-level `free()` does not throw, but
`std::free()` might.

In the end, a primitive `try`/`catch` seemed simpler and safer.