Experimental PQconnectdbParams support for C++20.
See #343.  This is not definite yet.  The new code will only be
activated if `PQXX_HAVE_CONCEPTS` is defined, and as yet, nothing will
define that automatically.

In principle this should work on any kind of input range of pairs,
tuples, or arrays (of size 2) of zero-terminated string types.  So that
includes `std::map<std::string, pqxx::zview>`,
`std::vector<char const *, char const *>`, and so on.

I created a few supporting concepts, and a bit of unifying API for known
types of zero-terminated strings: `char const *`, `char[]`,
`std::string`, and `pqxx::zview`.

Still untested are ranges of unknown size (would be easier to test with
coroutines I suppose), where the pre-allocation might fail to compile,
and `std::array` as either the range type or the pair type.