Support `to_string` etc. for `binarystring`.  (#312)
This will let you stream a `pqxx::binarystring` into a BYTEA field on
the database, or convert one to a string.  Even in an array.

Doing this was dubious before because binary escaping required the
connection object.  But now that all "reasonable" backend versions
support hex escaping, things get a lot simpler.

Haven't done the conversion in the other direction yet.  The problem
there is that backends might still decide to send the old format.