`check_cast`: Check for negative overflow.
The lower-end bounds check for negative values used
`std::numeric_traits<T>::min()`, but I wasn't aware that that has a
surprising definition for floating-point types: it returns not the
lowest finite number that the type can represent, as it does for
integral types, but the lowest _positive_ number.

Use the alternative `std::numeric_traits<T>::lowest()` instead.