test: Properly order tests executed by test runner (#354)
* test: Fix include's

Add missing string header include and move map header include.

 * test/runner.cxx: Add missing string and map header include's
 * test/test_helpers.hxx: Add missing string header include and remove unnecessary map header include

* test: Use std::string for test map keys

The order of tests run is determined by the test map order, and since some tests depend on the output from previous tests, the map must be ordered based on lexicographical comparison. Using `char const*` for the map keys results in the map order being determined by the pointer order, not the lexicographical string order. Additionally, the pointer order is indeterminate since the C-string storage and static initialization order is indeterminate[1]. Thus, use a std::string key to create a test map in lexicographical order and run the tests in the correct order.

 * test/runner.cxx: Use std::string for test map keys

[1] https://en.cppreference.com/w/cpp/language/initialization#Non-local_variables