Conversion: Support leading whitespace in integers.
Extends `from_string` to accept (some) leading whitespace.  The
documentation for composite types says that leading whitespace is
acceptable, depending on the conversion.  No details about when and
where it is acceptable, so I'm sort of flying blind here.  Anyway I'm
only planning to support the most basic example.

This is preparation for future composite types support.  I'm still
thinking about how to facilitate string conversions for these.
Unfortunately encoding could matter, so we may end up with a class
similar to the array parser.