More work on array/composite parsing.
Extracted more functions.  Turns out "double double-quote" escaping was
never (at least, according to the postgres docs) allowed in array
values, so that was not a bug.  However I did find that
backslash-escaping is allowed in an unquoted string.  The backend's
output won't make use of it, as far as I've seen, but technically it's
allowed.

I had fun generalising the scanning of unquoted strings for composite
types.  It depends on the context: in an array, you can have an
un-escaped closing parenthesis in an unquoted string, but a comma,
semicolon, or closing brace terminates the string.  In a composite
value, you can have an un-escaped closing brace in an unquoted string,
but a comma or closing parenthesis terminates the string.  Doesn't sound
like fun perhaps, but it's elegantly done using a parameter pack for a
compile-time variable number of characters, and a folding expression.