Use `std::size()`.
The new function replaces both container `std::size()`, and part of
`sizeof`: number of elements, but not byte size.  Most uses of `sizeof`
in libpqxx were on byte arrays, so those are gone now.

Once we have C++20, we'll be able to replace some casts here by using
`std::ssize()`.