Deprecate `namedclass`.
I hate this class.  It has its use, sure, but it brought in too much
complexity — notably virtual inheritance!

The libpqxx 7.0 changes may have simplified the situation a bit, and I
now see just two, distinct (non-overlapping) use-cases:
 1. The `transaction` hierarchy.
 2. `transactionfocus` and its children.

In both these cases, `namedclass` exists to check for, and report,
mistakes such as "you tried to open a transaction while another one was
active on the same connection."  So two things: common base class for
polymorphism, and a human-readable name for an object which may or may
not have been given an individual name.

Let's see how far we can get if we separate these, and streamline them
individually.

For starters, I'm not at all convinced that it's really useful to have a
specific class name in the case of the transaction hierarchy.  As far as
the user is concerned, they're all "transaction."  And if the developer
wants more detail, they'll get far more from giving the transactions
helpful names.  There goes the vtable.  Just a `name()` on the
transaction will do.

I haven't looked at `transactionfocus` in as much detail yet.  It's the
case where we do want polymorphism that isn't already there.  So yeah, I
fear the vtable stays.  But virtual inheritance?  No, that no longer
looks necessary.  Right now it's only there (far as I can see right now)
because `subtransaction` is both a `transactionfocus` and a
`transaction_base`, which is a `namedclass` in its own right.  But if I
tear these two things apart, well, who cares?

And... a `std::string` for the class name?  A `std::string_view` will do
just fine, thank you.  That string is hardly going to change!  Yes, yes,
I know: another option would be to construct the full description right
when we construct the `transactionfocus`.  But who wants to pay that
cost all the time for something we'll use so rarely?