Convert an entire `row` to a tuple in one go!
Adds a `row::as()` template.  Converts all of the row's fields at once,
returning them as a tuple of client-side objects similar to what the
newer streaming APIs do.

One thing I'd still love to do is to support `row::to()` taking a bunch
of references to variables for the respective fields.  But that will
conflict with the existing `row::to()` which takes a `std::tuple`, for
the case where there's only one column.  Not a problem right now,
admittedly, but in future I may want to support composite types
directly as conversions to tuples.  Which in turn will be a problem when
it comes to encodings, of course.  I'm seriously considering adding this
as a feature that works only with "safe" encodings such as UTF-8.