Cut transactions out of the `namedclass` hie4rarchy.
Of course it had to be more complicated than just removing the virtual
inheritance.  That special arrangement was also needed to enforce the
right initialisation order.

With this change, `transaction_base` and its descendants (except
`subtransaction`, which is _also_ a `transactionfocus`) are no longer
descendants of `namedclass`.  I've extracted the core of the `unique`
code into reusable functions.  Next I think I can fold `namedclass`
into `transactionfocus`, for a much simpler hierarchy.

I'm tempted to make `namedclass::classname` a virtual function, but I'll
need to investigage whether it's worth it.  It would save me the cost of
a `std::string` per object for the class name, but require a vtable.
Probably not what I want.  And bear in mind, we're dealing with very
small numbers of objects here.

Another thing I may choose to do later on is have an internal
"statement" class, completely invisible to the user, just so I can
make it a `transactionfocus` and re-use the machinery for detecting
and reporting clashes.