Make glyph scanners non-`noexcept`.
With older versions of clang and gcc, the sanitisers (in "audit mode")
complained about functions being called through an incorrect function
pointer type.  This appears to have happened because the glyph scanner
function for "monobyte" encodings was marked `noexcept`.  I guess that
affects the signature, yes.  Just annoying that it comes up at run time
and not during build.