Eliminate `do_abort` virtual call.
This member function can be called during destructors, so I replaced it
with a simple implementation in the base class, to be configured through
a constructor argument.

I left the function `virtual` though, in case there's anyone out there
who overrides it.  But there are no longer any overrides within libpqxx.

Also, I eliminated the "nascent" state in transactions.  We weren't
setting it anywhere.

This eliminates one or two sanitizer warnings in g++ 7 when compiling a
shared library.  Unfortunately the bulk of these warnings are still
there.