Factory for `stream_from` with `from_query`.
This is my way out of the quagmire of constructor overloads.  Instead of
constructing a `stream_from` and adding a `from_query` argument just to
disambiguate the call, you can now call `stream_from::query()`.

Obviously I will also want to do this for the `from_table` case.  But
I'd like to get those overloads right from the start, and that means
Concepts, which means C++20: `span`, `range`, and being able to treat
all iterator types differently from all string types in overloading
resolution.  Or perhaps for starters I'll find a nice way to tell the
overload for "one column, as a string" apart from the overload for
"run-time number of columns, as a container."  Or similar for "two
columns, as strings" versus "run-time number of columns, as a pair of
iterators."