Bump to 7.4.0.  Named stream constructors.
Replaces all constructors for `stream_to` and `stream_from` with factory
functions.  Deprecates all the poorly differentiated old constructors,
as well as the more recent, still awkward, constructors differentiated
by marker types.

This also enables a code path where the user can reuse some quoting (or
even hard-code it) between invocations.

Introduces a new type alias, `table_path`, for naming a table.  A simple
string was no longer tenable because we couldn't tell a table name
containing a dot from a table name qualified by a schema name.

The new scheme is not perfect, but I think it strikes a nice balance
between flexibility and simplicity.  There are not a lot of
constructors, but if you absolutely have to have a dynamic columns list,
that's supported in a uniform way.  There is minimal code duplication
between the stream types.