Restrict `fill_buffer` to non-string types.
Well, sort of.  We don't support `char` as a C++ type that can be
written to the database directly, since it's ambiguous.  (Does it mean
"a single character," or "a single-byte character," or "a one-byte
integer," or "a one-byte _unsigned_ integer?)  So, block the overload in
the case where the container being passed has `char` as its content
type.  Such a type would presumably be meant as a string.