Stop "aborting" nontransactions on closing.
One of the `nontransaction` constructors failed to pass the null
`abort_cmd`, meaning that if the transaction was destroyed without
having been committed or aborted, it would send a "ROLLBACK".

This didn't really do any harm, since a nontransaction can't roll
back, but it wasted performance and logged warnings.

Fixes #419.