Drop legacy tests 31 & 49 for clang C++20 bug.
This seems to be a very rare bug in clang, with nary a mention on the
Internet, but I couldn't find a decent way to work around it.  When
these tests do string comparison on result fields, the compiler
transforms the code to use `operator<=>`... and then complains that its
own transformed code uses a zero constant (`0`) to represent the null
pointer.

Luckily, these are _legacy_ tests and removing them did not upset any of
the other tests.