Avoid deprecated `stream_from` ctor.
Couldn't get this right with `std::make_unique()`, which just won't work
with a named constructor.  Funny little hole in the standard, really.

But I really wanted to get rid of this deprecated call, because Visual
Studio felt that the call was being made from within `make_unique` and
so there was no way to suppress the deprecation warning on MSVC.