Tweak `[[likely]]` check for gcc 8.
Fixes #452.

The feature check for `[[likely]]` and `[[unlikely]]` was passing on
gcc 8, albeit with a warning.  But actual library code using those
attributes did not compile at all.

Turned out the way the check was formulated happened to hit a rare case
where gcc 8 issues a warning, not an error.