Add `stdc++fs` lib check for CMake.
Both the autotools build and the CMake build now check whether they need
to link in the `stdc++fs` library.  Seems to be needed specifically in
the case of gcc 8, which still seems to be the current standard in
Debian stable.

Annoyingly, the test still crashes for me when building with gcc 8 and
the autotools build.  The `std::filesystem::path` destructor calls the
destructor of its internal `std::vector` of path components, which
segfaults.

If Debian stable didn't still seem to rely on gcc 8, I'd drop support
right here.  But let's give it a bit more time and consultation.  Ideas,
anyone?