Deprecate "result slicing."
There is a very old feature that lets you view a `row` as a new `row`
containing only a sub-range of its fields.  Never seemed particularly
useful or complete, and there's probably better ways of doing it now.

The fact that it was still documented as "a new feature" shows that
there hasn't been much talk or work related to it since... when?  It
feels like 2005.  So it's time to YAGNI that thing and, as soon as we
can get away with it, smooth out the annoying little complications it
adds to the code.  Should also take a tiny bit of code out of most
applications' inner loops, so expect a tiny speed boost.  Probably not
measurable, but it'll _feel_ good.