Work around gcc 8 crash in `path` destructor.
On gcc 8 on Ubuntu, the test for passing a `std::filesystem::path` to
`pqxx::blob` was crashing in the `path` destructor.  Looks like a bug
in Ubuntu's version of the gcc-8 implementation of the filesystem
library.  I think the same tests actually ran fine with gcc-8 on
Debian.

Had we had a configure/cmake check for `std::filesystem` support, I
could have built this extra check in there.  But we don't, and it's
not worth the overhead.  Especially so since the problem only happens
at run time: we'd have to write a check, and have both `configure`
and the CMake build actually _run_ the check, not just compile and
link it.  Which adds concerns for cross-compiling situations, where
the host platform is not necessarily able to run the code that the
build produces.

Also, I don't know the extent of the problem.  It could be fine in
some other situation, even with gcc-8 on Ubuntu.  So I'm simply
disabling the test on gcc-8.  If anybody else hits the same problem, it
will show up as happening in their own code, where they can debug it.