Only use `[[likely]]` in C++20.
This attribute (and its negation, `[[unlikely]]`) were only defined in
C++20, but I previously enabled them whenever the compiler allows them.
After all, compilers might implement these attributes a bit more
aggressively.

But the warnings got so annoying, and the hope of benefits is so minor,
that I'm disabling this in C++17.