Fix `stream_to` encoding.  Bump to 7.7.0.
For #473, stop octal-encoding any non-ASCII characters in `stream_to`.
The server interprets the numeric encoding as being in its _database_
encoding, not the _client_ encoding.

Instead, make the loop encoding-aware so that it can escape just the
few special characters that it really needs to escape, but without
accidentally doing so for the same byte values if they happen to occur
in the middle of a multibyte character.

Unfortunately this requires an ABI-incompatible change: `stream_to`
needs to "remember" the encoding.  Hence the bump to 7.7.0.