Clearer error for ambiguous `char` conversion.
Fixes #481.

There is no string conversion for `char`, `signed char`, and `unsigned char`.
For me, those conversions would just be too ambiguous to be safe: should we
convert them as very short strings, or as very small integers?

The one problem was that when application code does try to convert one of
these types (e.g. because the compiler typedefs `int8_t` as `signed char`!)
it fails with a very obscure error message.  People interpret it as a
missing specialisation.

So, ensure that it at least fails with an error that mentions the root
problem: there was an attempt to perform an ambiguous, and disallowed,
conversion.