Support nonblocking connection.
See #487.  We used to have a `connection` class _hierarchy,_ and one of
those did something similar.  But this new class is much more limited in
scope and keeps things simple.

In the future hopefully this can form a basis for C++20 async support.
But for now, any application with a wait loop based on `poll()` or
`select()` should be able to use it.