Replace `std::isdigit` with a safer alternative.
Turns out I was probably using `isdigit` incorrectly (it lives in `std`,
according to cppreference), but moreover, there's a pitfall: if you pass
it a negative number (as might easily happen on a system where `char` is
signed), the results are undefined.  So proper usage involves _casting_
the argument yourself.  Insane.