Use `short` when rendering year as string.
The C++ standard says a `std::chrono::year` must lie between -32767 and 32767
inclusive, i.e. it must fit in a 16-bit signed integer.  A `short` must be at
least 16 bits wide, and can be narrower (but not wider0 than an `int`.

So, representing an outgoing year as a `short` may tighten the buffer budget
needed for the conversion.  We don't need the same narrowing while parsing;
it's not likely to save any time, and in the event that a year is out of the
permitted range, using an `int` lets the `year` parser rather than the number
parser generate the error, resulting in a clearer, more specific error
message.