Don't support `year` etc. conversions.
I've gone back and forth on this a few times.  I now do _not_ think
there should be separate string conversions for `std::chrono::year`,
`std::chrono::month`, and `std::chrono::day`.

The main reason for this is that there's no way to re-use a `year`
conversion as part of a `year_month_day` conversion.  The problem is
the "BC" suffix, which isn't anywhere near adjacent to the year
itself.  Makes it very hard to handle the year in isolation.

Several other arguments:
1. There are no corresponding SQL types that I can see, anyway.
2. It'd be nice to support years as SQL integers, but they would not
   follow the same format as years in SQL dates.
3. Would an isolated year value be "BC/AD", or "Common Era"?

So, a string conversion for `year` would be an invitation to various
bugs and ambiguities.  So best not to offer one.  And if we can't have a
conversion for `year`, what sense does it make to have ones for `month`
and `day`?