Use `std::cmp_greater` etc. if available.
This is a C++20 feature.  It may not seem like a big deal, but it lets
us avoid some nasty casts (and their concomitant range checks).

Now if only these functions supported floating-point types as well, we
could similarly update `check_cast<>()`.  But we can't have it all I
suppose.