Introduce & use new internal funcion `find_char`.
This new function template scans a string (following encoding rules for
character boundaries) and looks for the first occurrance of any of a
given set of ASCII characters.

It will enable me to rewrite most or all of the loops that do
encoding-specific scanning of strings.  For starters it will simplify
and hopefully speed up the inner loops.  But also, in the future, it
will allow me to replace full-on encoding-specific scanning of character
boundaries with faster, dumber search loops, for encodings like UTF-8
where a given ASCII character can never occur as a byte value in a
different, multi-byte character.

This in turn will mean that a search through a UTF-8 string won't have
to skip from one UTF-8 character boundary to the next.  Instead, we can
just do a minimal straight-through loop that reads and checks one byte
at a time.

But that's all in the future.  For now, I just replaced one of the loops
using the new function, and also dropped another encoding-specific
search function which we weren't using.