Disable `std::filesystem::path` support on Windows.
Where supported, we have some functions that accept a `path` but
forward the path string to an equivalent call based on C strings.

On Windows, however, `std::filesystem::path::c_str()` returns not a
`char` string but a `wchar_t` string.  And since there's no overload for
that, the string gets promoted back to a `path`.