Try to fix readthedocs build with ancient compiler.
I just changed `configure.ac` to _require_ C++17 but not _try to get_
C++17.  It's up to the user to provide options as needed to make C++17
happen.

But the readthedocs build uses some ancient environment, which does not
give us C++17 by default.  Try to set `CXXFLAGS` in the documentation
configuration.

And while we're at it, set `-O0` to speed things up.