Rework variables.
This part was always messy.  The development of variables support in libpqxx
revealed some warts in how PostgreSQL handled variables at the time.  They
behaved transactionally in some scenarios, but not in others.  I believe this
was cleared up later, with a clear distinction between local variables and
session variables.  If memory serves, I updated the implementation in libpqxx
but did not re-work the design.

Now it's time to do that.  It closes off a possible route to SQL injections,
and reduces complexity by dropping support for local variables.  Session
variables rightly belong on the `connection`, and supporting them there
offers real convenience.  (To set a session variable outside a transaction
you might need to create a `nontransaction` otherwise.)  But a local variable
is about as easy to set through SQL as it is with a dedicated function.

Also, it looks as if a variable can't be null.  So I explicitly forbade
attempts to set one to null, with a specific new execption type.