Extract wait functions into new module.
Finally got sick of managing all those conditional includes for `poll()`
and `select()`.  Turns out I'd missed one of the alternatives in one of
the call sites.  It's a lot of pollution at the top of the source file,
too.

This may take some time to debug on various systems.  Starting to wonder
whether it might be worth doing generalised support for `epoll()` and
kqueue as well.