Fix a bug that unexpectedly created library symlink (#550)
* Fix a bug that unexpectedly created library symlink

There was a bug in CMake that created library symlink when the CMake target name and the generated library name were different, but the library symlink was created when the two names were the same.
This bug occurred when the generated library name (`OUTPUT_NAME` property) contained a CMake generator expression (e.g. `OUTPUT_NAME=$<IF:$<PLATFORM_ID:Windows>,pqxx,pqxx-7.7>`).
If the library symlink is created when two names are the same, the source and destination of the link will have the same name, resulting in a circular reference (e.g. `libpqxx.a -> libpqxx.a`).
Avoid unexpected creation of library symlink by comparing the CMake target name with the generated library name when evaluating CMake generator expressions.

* Fix order of evaluation of CMake variables