Callback-based `for_each` methods. (#528)
One new function, `pqxx::result::for_each`, applies a callback you give it to each row in the result set.  The neat thing is that it automatically converts the fields to the types required by the callback's corresponding parameter.  (Of course this won't work for `auto` parameters, because in that case, the callback and `for_each` will both be expecting the other to specify a type.)

Similarly, `pqxx::transaction_base::for_each` executes a streaming query and applies your callback to each row as it comes in.

Thanks @KayEss for helpful feedback.