Fix broken nonblocking connection on Windows.
I finally got the tests running in Windows on appveyor.  Which revealed
a hanging test in Windows, for the nonblocking connection class.

The `pqxx::connecting` class tried to manage blocking/nonblocking states
on the connection.  But there does not seem to be any need: libpq does
all of that for us.