Get Windows tests working & passing. (#560)
I added an appveyor.yml. This should mean I no longer have to configure appveyor in the web UI. And more of the test is set up automatically, without awkward scripting in various convoluted languages.

But it also gave me the opportunity to get the tests running on Windows. And guess what? They didn't all pass! The test for asynchronous connection (using pqxx::connecting) would hang. Turns out it was completely unnecessary for libpqxx to manage blocking/nonblocking mode on the socket. So that's fixed now.