Support querying & converting 1 row of data.
This makes it convenient to query either exactly 1 row of data (using
`query1()`), or at most 1 row of data (using `query01()`), and convert
the results straight to the desired client-side types.

The version that allows an empty result returns a `std::optional`, which
obviously is empty if the query produced no data.