Markdown lint (#581)
Trying [mdl](https://github.com/markdownlint/markdownlint/) on the Markdown files in libpqxx.  Had to tweak a few settings and suppress a few warnings, and the blank line above a list will take some getting used to.  But otherwise, okay, I'll do it the way the tools want me to.