Make `stream_to` accept `optional<string_view>`. (#597)
Fixes #596.

It'd be nice to support each of `optional`, `shared_ptr`, `unique_ptr`,
and potentially other smart pointer types for each of these string
types.  But doing that without C++20's "concepts" is going to be a
little hairy.

Also switching CircleCI tests to use clang, just because the `debian/unstable` image keeps failing with a warning in gcc's own headers.  Can't reproduce that on my own machine, so I hope it's just a transient that's already fixed in Debian.