Bump minimum libpq/postgres to version 10.
The PostgreSQL project no longer supports postgres 9, so I'm dropping
it off the libpqxx support list as well.

The only consequence code-wise is that libpqxx will now assume that
libpq has a `PQencryptPasswordConn()`.

Actually might be nice to have a compile-time check for an acceptable
libpq version.  But that's a separate project.