Specialise array parser per encoding.
More inlining of encoding-specific inner-loop code.  This should be
faster, but also, it's a step leading up to support for different
separators in arrays (based on the element type).

There's loads more work that I'd like to do on array parsing.  For
instance, automatic conversion to a client-side element type.  And it
may well end up being more efficient as well, since I won't need to
allocate a new `std::string` for every element.

Unfortunately a total overhaul will probably have to wait for
`std::mdspan` and such to set a canonical style for this kind of
range.