Extract "encoding group mapping for ASCII search."
We've got a set of encoding groups, each with its own algorithm for
finding the character boundaries.  But for purposes of finding a
particular ASCII character, most of these look a lot like monobyte
encodings.  For example, you can just scan a UTF-8 string for your
desired ASCII byte value, and completely ignore the existence of
multibyte characters.  (In fact, this was one of UTF-8's main design
objectives.)

I just extracted a function to map encoding groups: "when searching
for a particular ASCII character in encoding group X, treat it as if
it were encoding group Y."  In practice so far, the dichotomy is always
just "treat as monobyte" versus "no, I need to figure out the character
boundaries first."