Link statically.
This sometimes exposes failure to export an object from the library.
But also, in this one case, I'm getting a weird link error otherwise:
it says I'm not linking to `asan` as the first library in the linker
command line.

Which is weird because (1) it doesn't happen on my own machine with
more or less the same OS, compiler, and options; and (2) linking this
particular library is supposed to be handled by the compiler itself
when we enable address sanitisation.