WIP: Get Github's new code checking working (#616)
Github suddenly replaced LGTM.com with CodeQL.  It generated a default config for that, which for some reason wasn't working.  It didn't find either of the build systems (CMake and configure, pretty standard!) so I had to configure a manual build instead.

It's probably not the best way to do things — I had to make sense of fairly poor documentation in a hurry.  For one, this setup seems to build the C++ code twice: once in order to do static analysis on the C++ code (not clear to me why that's needed) and then again in order to do static analysis on the Python code.  Huh?

CodeQL produced some code warnings.  There was one warning about a function marked `[[maybe-unused]]`... _being unused._  Gee, thanks.  Oh well, I think I've fixed that and improved the code in the process, but being able to write that "maybe unused" function without triggering a warning was actually pretty useful in the code's evolution.

The other warnings were false positives generated by code snippets that the configure script generates.  It may well be arrogant of me, but besides the warnings being inappropriate in this case, I am not impressed with the "improvements" they suggested.  Is "foo" really too short a variable name?  It all depends.  But certainly the suggested replacement style is totally unacceptable: too long, and in dromedaryCase which I now think is an abomination.  I miss LGTM.com already.